Changed FLA to export SWC, and more
Made some modifcations to the Publish Settings of browser.fla. Before,
the ClassAction browser was compiled in Flash CS4 using the FLA. Now, it
is being compiled from the command line starting with Main.as. There are
still necessary resources (graphics) inside the FLA, so now they are
being exported into a SWC which then gets folded in by the compiler.
This change in arrangement required some changes to class inheritance in
order to work (specifically, HeaderBar and SearchPanel).

Other changes:
- ClickableText had to be modified since ContextMenu is NativeMenu in
the AIR app.
- Modified the Downloader base path init in Main.as.