Fix (dubious) for unusual bug
For reasons I don't understand, the runtime started failing at
initialization. The error was #1034, could not convert
flash.display::MovieClip@--- to fl.controls.Button. But looking at the
stack trace showed SearchPanel.as line 59, which had the declaration for
_panesWidth. Changing this to a hard-coded value made this runtime error
disappear. It seems like a race condition bug, in which case I'm not
confident that the bug was fixed.